lcatricity_dashboard/requirements.txt
httpx
pandas
sqlalchemy
python-dotenv
psycopg2-binary
streamlit
matplotlib~=3.9.1
